Create a remote author fabricator with a remote url and a domain that matches
I got confused because it's possible to have an author with a remote url that has a different domain than the author's domain, but it's not very likely in the real world. So I made a remote author fabricator for use in tests that sets both of them more realistically.