Adds a checkbox that might make it a bit annoying to post spam on rstat.us
When javascript is enabled, the checkbox will get checked and hidden, so legitimate users won't experience anything different.

If a legitimate user has javascript disabled, they'll see an unchecked checkbox that says "I am not a spammer", and if they check it, their update will be posted.

Any request that does not have this checkbox checked will not result in a posted update.

This might stop spam for a bit, until they figure out they need to send something in this parameter...