Configure Travis to run with multiple ruby versions
In commit 65f0f67a, I broke the build for 1.9.2 on Travis by specifying 1.9.3 in the Gemfile. This commit adds a different Gemfile without the ruby version for use by Travis (which means we'll need to update both in lockstep, not ideal, I know).

This also adds ruby 2.0.0 to our Travis matrix but allows it to fail for now.